requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pyyaml>=6.0
httpx>=0.26.0
piper-tts>=1.2.0
numpy>=1.24.0
